+++ src/rest/defaults.ts
@@ -1,13 +1,14 @@
 import {Type} from '@angular/core';
 
 import {RestMiddleware} from './rest.interface';
-import {ReportProgressMiddleware} from './middlewares';
+import {ReportProgressMiddleware, ResponseTypeMiddleware} from './middlewares';
 
 /**
  * Definition of basic default array of rest middlewares order
  */
 export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
 [
+    ResponseTypeMiddleware,
     ReportProgressMiddleware
 ];
 
@@ -16,5 +17,6 @@ export const BASIC_DEFAULT_REST_MIDDLEWARES_ORDER: Type<RestMiddleware>[] =
  */
 export const BASIC_DEFAULT_REST_METHOD_MIDDLEWARES: Type<RestMiddleware>[] =
 [
+    ResponseTypeMiddleware,
     ReportProgressMiddleware
 ];