+++ src/rest/middlewares/index.ts
@@ -1,2 +1,3 @@
 // export * from './parameters.middleware';
 export * from './reportProgress.middleware';
+export * from './responseType.middleware';